2.1.27
Global saltkey change implemented

+++ sources/main.functions.php
@@ -1651,7 +1651,7 @@ function encrypt_or_decrypt_file($image_code, $image_status, $opts) {
             );
         }
     } elseif (isset($_SESSION['settings']['enable_attachment_encryption']) && $_SESSION['settings']['enable_attachment_encryption'] === "0" && isset($image_status) && $image_status === "encrypted") {
-        // file needs to be encrypted
+        // file needs to be decrypted
         if (file_exists($_SESSION['settings']['path_to_upload_folder'].'/'.$image_code)) {
             // make a copy of file
             if (!copy(
@@ -1709,11 +1709,14 @@ function debugTeampass($text) {
 * DELETE the file with expected command depending on server type
 */
 function fileDelete($file) {
-    // define if we under Windows
-    if (strpos(dirname(__FILE__), '/', 0) !== false) {
-        unlink($file);
-    } else {
-        $lines = array();
-        exec("DEL /F/Q \"" . $file . "\"", $lines, $deleteError);
+    if (is_file($file)) {
+        @close($file);
+        // define if we under Windows
+        if (strpos(dirname(__FILE__), '/', 0) !== false) {
+            unlink($file);
+        } else {
+            $lines = array();
+            exec("DEL /F/Q \"" . $file . "\"", $lines, $deleteError);
+        }
     }
 }
